exact_requirements.txt
absl-py==2.3.1
accelerate==0.21.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
backports.tarfile==1.2.0
basicsr==1.4.2
Brotli==1.1.0
brotlicffi 
build==1.3.0
CacheControl==0.14.3
certifi
cffi
charset-normalizer==3.4.2
cleo==2.1.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
crashtest==0.4.1
cycler==0.12.1
decorator==4.4.2
distlib==0.4.0
distro==1.9.0
dulwich==0.22.8
exceptiongroup==1.3.0
facexlib==0.3.0
fastjsonschema==2.21.1
filelock==3.18.0
filterpy==1.4.5
findpython==0.6.3
fonttools==4.59.0
frozenlist==1.6.0
fsspec==2025.7.0
future==1.0.0
g4f==0.5.8.0
gfpgan==1.3.8
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.25.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
Jinja2==3.1.6
jiter==0.10.0
keyring==25.6.0
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.7.3
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.9.4
mkl-service==2.4.0
mkl_fft
mkl_random
more-itertools==10.7.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.4.4
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==1.23.5
nvidia-ml-py==12.575.51
openai==1.98.0
opencv-python==4.11.0.86
packaging==25.0
pbs-installer==2025.8.8
pillow==11.3.0
pillow_heif==1.1.0
pip-tools==7.5.0
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry-core==2.1.3
proglog==0.1.12
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
pycparserpycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
pynvml==12.0.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
PySocks
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.13.0
realesrgan==0.3.0
regex==2025.7.34
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-image==0.24.0
scipy==1.13.1
shellingham==1.5.4
shiboken6==6.9.1
six==1.17.0
sk-video==1.1.10
sniffio==1.3.1
tb-nightly==2.21.0a20250729
tensorboard-data-server==0.7.2
tifffile==2024.8.30
tokenizers==0.13.3
tomli==2.2.1
tomlkit==0.13.3
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.67.1
transformers==4.32.0
trove-classifiers==2025.8.6.13
typing-inspection==0.4.1
typing_extensions
urllib3 
virtualenv==20.32.0
Werkzeug==3.1.3
win-inet-pton
yapf==0.43.0
yarl==1.20.0
zipp==3.23.0
zstandard==0.23.0
